style(clippy): clear all 44 lints in eldenring-archipelago
Clippy had never run on this crate (Actions was disabled on the fork), so this is the whole
backlog. No behaviour change -- every fix is a rewrite into the idiom clippy names.

  12  map_or(false, ..) -> is_some_and(..)   -- ALL IN contract_gen.rs, WHICH IS GENERATED.
                                               Fixed in the EMITTER (eldenring/contract.py),
                                               not the output. rustfmt honours '@generated';
                                               CLIPPY DOES NOT -- a generated file is linted
                                               like any other, so the emitted Rust has to be
                                               lint-clean at the source or the next regen
                                               silently undoes the fix.
  11  doc_lazy_continuation                 -- markdown misreads, not prose problems: a doc
                                               line starting with '+' IS a bullet, and prose
                                               butted against a list reads as a continuation
                                               of the last item. A blank '///' ends a list.
   7  collapsible_if                        -- edition-2024 let-chains (core.rs already used
                                               them).
   3  type_complexity                       -- named the slot_data-shaped tables (DropSlot /
                                               RollTable / StockRow / StockTable / FmgInjects).
   2  question_mark, 2 redundant_closure, 2 unnecessary_filter_map ( in a
   2  manual_is_multiple_of, 1 manual_range_contains, 1 unnecessary_cast
   1  empty_line_after_doc_comments         -- an orphaned '///' block documenting
                                               decode_weapon_id, a fn that was DELETED.
                                               Demoted to '//'; the note is still useful.

Verified before pushing: rustfmt clean, and all 14 touched files pass
'rustc -Zparse-crate-root-only --edition 2024'. That caught 3 orphan closing braces I left
behind while collapsing if-chains -- collapsing two ifs removes a BLOCK, so a '}' has to go
with it. (A real cargo check still does not fit in the sandbox; CI remains the gate.)

Clippy stays continue-on-error in THIS commit so CI can adjudicate. Re-armed as a hard gate
in the follow-up, once green.

+++ crates/eldenring-archipelago/src/upgrades.rs
@@ -178,11 +178,12 @@ pub fn apply_auto_upgrade(full_id: i32) -> i32 {
 /// ER category nibble mask / weapon-category constant (er_codec mirror; weapons are category 0x0).
 const ROW_ID_MASK: u32 = er_codec::ROW_ID_MASK;
 
-/// Pure id-math decode: returns (base_row_id, level) for an upgradeable WEAPON FullID, else None.
-/// base = row - row%100; level = row%100. Category guard mirrors C++ `WeaponInfo`:
-///   `(uint32(itemId) & CATEGORY_MASK) != CATEGORY_WEAPON` rejects non-weapons; row range
-///   `[1_000_000, 90_000_000)` skips system/NPC ids. Weapons are er_codec::CATEGORY_WEAPON (0x0).
-// `decode_weapon_id` was a byte-for-byte copy of er_logic::upgrades::decode_weapon_id. Production never
+// (Historical, kept as a note.) `decode_weapon_id` was a pure id-math decode: (base_row_id, level) for
+// an upgradeable WEAPON FullID, else None. base = row - row%100; level = row%100. Category guard mirrored
+// C++ `WeaponInfo`: `(uint32(itemId) & CATEGORY_MASK) != CATEGORY_WEAPON` rejects non-weapons; row range
+// `[1_000_000, 90_000_000)` skips system/NPC ids. Weapons are er_codec::CATEGORY_WEAPON (0x0).
+//
+// It was a byte-for-byte copy of er_logic::upgrades::decode_weapon_id. Production never
 // called it -- apply_auto_upgrade delegates to er-logic, which uses ITS copy -- so this one existed only
 // to be unit-tested. A tested copy that production does not call is the exact test/prod drift the replay
 // tier exists to kill (CONTRIBUTING: "a green predicate with no production caller is a spec, not a fix").
@@ -208,7 +209,7 @@ pub(crate) fn weapon_track_and_cap(base: i32) -> Option<(i32, bool)> {
     }
     // TRACK from materialSetId, cap from the run above -- host-tested predicate (owns the somber
     // clamp + the not-upgradeable guard).
-    er_logic::upgrades::classify_track(k - 1, weapon.material_set_id() as i32)
+    er_logic::upgrades::classify_track(k - 1, weapon.material_set_id())
 }
 
 /// RE-A3 RESOLVED (typed binding): highest +N currently held on the given smithing track
